docker-flask/app/requirements.txt
flask
python-dotenv
flask-restful
ibm_db
pyjwt
flask-expects-json